warehouse_inventory/requirements.txt
Django==4.2.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0
